+++ src/utils/common/key.ts
@@ -1,4 +1,4 @@
-import { B64EncryptionResult } from 'utils/crypto';
+import { B64EncryptionResult } from 'types/crypto';
 import CryptoWorker from 'utils/crypto';
 import { getData, LS_KEYS } from 'utils/storage/localStorage';
 import { getKey, SESSION_KEYS } from 'utils/storage/sessionStorage';
